frontend: rename Confirm to Confirm-Legacy
The Confirm.tsx is deprecated and should not be used in new code.
Renaming it makes that clear, and lets us create a new Confirm
component.
